One sided comparison for some base elements and Mac Lane elements
This covers a common use case. Of course, it only handles the case when the
elements are not equal. At the moment we try not to push the precision of the
approximation of the limit valuation so after a few computations with the
valuations, a NotImplementedError might turn into a False.